add TruncateSoftwareUpgradePlan
